backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
anthropic>=0.41.0
openai>=1.57.4
langgraph==0.2.60
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-openai==0.2.14
langchain-qdrant==0.2.0
qdrant-client==1.12.1
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.0.1
